% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCXst.R
\name{addCXst}
\alias{addCXst}
\title{addCXst}
\usage{
addCXst(epi.melt, epi.original, Cst = "Cst", Xst = "Xst", id.original = "kid")
}
\arguments{
\item{epi.melt}{tabla  con secuencia de episodios generada por meltEpi con clave "kid"
para enlazar con episodio originales}

\item{epi.original}{tabla con informacion de episodio originales (campo ID, Cst y Xst)}

\item{Cst}{vector con la informaci칩n del estado actual}

\item{Xst}{vector con la informaci칩n del  estados de salida}

\item{id.original}{cadena de caracter con nombre de columna  con clave identificadora
de episodio}
}
\value{
data.table de subepisodios ampliada con dos nuevos campos
        Cst (Current Status) u Xst (exit status) tomado dela tabla
        original 'epi.original'
}
\description{
A침ade informacion sobre estado actual (al comienzo del episodio:
Cst) y estado a la salida de episodio (Xst) a los subepisodios
particionados, heredados de otra data.table con la la inforamacion
de Cst y Xst de episodios originales no particonados pero
identificados por clave de episodio
}
\details{
Cada fila de id.original es un episodios con estados constante =
'Cst", al terminar el episodio se produce,o no, una transici칩n de estados a 'Xst'
}
\examples{
\donttest{
epi.original <- data.table::data.table(
 kid = 1:3 ,
 start.times  = c(16,21,32)     ,
 durations  = c(12.5,19.4, 7.3) ,
 Cst = c(0,0,0),
 Xst = c(1,1,1)
 )

chop (start.times =  start.times,
     durations   = durations,
     breaks = seq(10,50, by=5),
     timedim = 'edad', data = epi.original)  -> epi.seq

meltEpi(epi.seq) -> epi.sub

addCXst(epi.sub, epi.original,
          Cst= 'Cst',  Xst='Xst',
          id.original = 'kid') -> epi.sub2
epi.sub2
  }
}
