% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calTime.R
\name{cal.age2}
\alias{cal.age2}
\title{cal.age2}
\usage{
cal.age2(date.start, date.end, decimal.precision = 3)
}
\arguments{
\item{date.start}{fechas de inicio (nacimiento habitualmente)}

\item{date.end}{fecha en la que se calcula la edad exacta}

\item{decimal.precision}{decimales de aprosimacion de año}
}
\description{
Calcula edad exacta en años decimales.
}
