% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine2c.R
\name{combine2c}
\alias{combine2c}
\title{combine2c}
\usage{
combine2c(epi.seq1, epi.seq2, dec.precision = 3)
}
\arguments{
\item{epi.seq1}{secuencia de subepisodio de un episodio dado
definido en una dimension temporal dada (ejemplo edad)}

\item{epi.seq2}{secuencia de subepisodio del mimos episodio anterio
pero definido en segunda dimension temporal dada (ejemplo
tiempo calendario)}

\item{dec.precision}{maxima precion decimal para comparaciones
puntuales}
}
\value{
data,table con los registros inteseccion de las tabla de episodios y
        y una nueva variable \emph{duration.#...} con la intesercion de los
        subepisodios en cada una de las dos escala temporales de los
        ficheros fuentes.
}
\description{
Combina dos objeto del tipo "secuencia de subepisodios"  producido sobre una
misma tabla madre de episodios producido por la función  por
\code{chopepi::chop}  pero particionado en dos dimensiones
temporales distintas. Ejemplo años cumplidos y anualidades
}
\examples{
\donttest{
## Trocea una tabla de episodios en 3 escalas termporales distintas
  start.times <- c(16,21,32)
  durations <- c(12.5,19.4, 7.3)      ; breaks <- seq(10,50, by=5)
  start.times2 <- c(1997,2005.2,2007) ; breaks2 <- seq(1990,2020, by=5)
  start.times3 <- c(3.2,4,2.3)        ; breaks3 <- seq(0,25, by=5)

  chop (start.times,  durations,  breaks,   timedim = 'edad' )   -> epi.seq1
  chop (start.times2, durations,  breaks2 , timedim = 'periodo') -> epi.seq2
  chop (start.times3, durations,  breaks3 , timedim = 'duraMat') -> epi.seq3

## Combina las tres escalas en dos pasos ...
  combine2c (epi.seq1,epi.seq2) -> epi.1y2
  names ( attributes(epi.1y2) )
  combine2c (epi.1y2,epi.seq3) -> epi.1y2y3
  names ( attributes(epi.1y2y3) )
  epi.1y2y3
}
}
