% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meltEpi.R
\name{meltEpi}
\alias{meltEpi}
\title{meltEpi}
\usage{
meltEpi(
  epi.breaks,
  type = c("factor", "left", "middle", "right", "integer"),
  dec.precision = 3
)
}
\arguments{
\item{epi.breaks}{objeto episodio particionado, al menos por una
dimensión temporal}

\item{type}{tipo de etiqueta asignada a cada banda dimensional}

\item{dec.precision}{escala de precion para las comparaciones}
}
\value{
un data.table donde cada registro de episodio se ha divido en tamtos
        subepisodios como elemetos haya en el vectos de troceado generado
        prevaimente por la combinacion de funciones \code{chop} y
        \code{combine2c}
}
\description{
Toma una tabla de episodios  (durations) preprocesado por \code{chop} y que
cada fila contiene una lista con el vector de subepisodio troceado por una o
mas escalas temporales, y genera un registro individual por cada uno de los
subepisodio contenidos en el episodio.
}
\examples{

\donttest{
## Trocea una tabla de episodios en 3 escalas termporales distintas
  start.times <- c(16,21,32)
  durations <- c(12.5,19.4, 7.3)      ; breaks <- seq(10,50, by=5)
  start.times2 <- c(1997,2005.2,2007) ; breaks2 <- seq(1990,2020, by=5)
  chop (start.times,  durations,  breaks,   timedim = 'edad' )   -> epi.seq1
  chop (start.times2, durations,  breaks2 , timedim = 'periodo') -> epi.seq2

## Combina las dos escalas en dos pasos ...
  combine2c (epi.seq1,epi.seq2) -> epi.1y2
  meltEpi(epi.1y2) -> epi.melt
}

}
