% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine2c.R
\name{combine2c}
\alias{combine2c}
\title{combine2c}
\usage{
combine2c(epi.seq1, epi.seq2, dec.precision = 3)
}
\arguments{
\item{epi.seq1}{secuencia de subepisodio de un episodio dado
definido en una dimension temporal dada (ejemplo edad)}

\item{epi.seq2}{secuencia de subepisodio del mimos episodio anterio
pero definido en segunda dimension temporal dada (ejemplo
tiempo calendario)}

\item{dec.precision}{maxima precion decimal para comparaciones
puntuales}
}
\value{
data,table con los registros inteseccion de las tabla de episodios y
        y una nueva variable \emph{durations.#...} con la intesercion de los
        subepisodios en cada una de las dos escala temporales de los
        ficheros fuentes.
}
\description{
Combina dos objeto del tipo "secuencia de subepisodios"  producido sobre una
misma tabla madre de episodios producido por la función  por
\code{chopepi::chop}  pero particionado en dos dimensiones
temporales distintas. Ejemplo años cumplidos y anualidades
}
\examples{
\donttest{
epi.original <- data.table::data.table(
 kid = 1:3 ,
 start.times1  = c(16,21,32)     ,
 durations  = c(12.5,19.4, 7.3) ,
 start.times2 = c(1997,2005.2,2007),
 Cst = c(0,0,0),
 Xst = c(1,1,1)
 )

chop (start.times =  start.times1,
     durations   = durations,
     breaks = seq(10,50, by=5),
     timedim = 'edad', data = epi.original)  -> epi.seq1

chop (start.times = start.times2,
     durations    = durations,
     breaks       = seq(10,50, by=5),
     timedim      = 'edad', data = epi.original)  -> epi.seq2

combine2c (epi.seq1,epi.seq2, dec.precision = 2) -> epi.1y2
names (epi.1y2)

}
}
